requirements.txt
crewai>=0.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0
google-generativeai>=0.3.0
langchain-text-splitters>=0.0.1
pydantic>=2.0.0
llama-index>=0.9.0
llama-index-llms-google>=0.1.0
llama-index-readers-file>=0.1.0
pypdf>=3.15.0
flask>=2.0.0
werkzeug>=2.0.0
gunicorn>=20.1.0